+++ app/src/main/java/com/example/scindapsus/model/Scene.java
@@ -7,18 +7,18 @@ import java.util.List;
  */
 
 public class Scene {
-    private int id;
+    private long id;
 
     private List<Line> lines ;
 
     private String name;
 
-    private int ordinal;
+    private long ordinal;
 
-    public void setId(int id){
+    public void setId(long id){
         this.id = id;
     }
-    public int getId(){
+    public long getId(){
         return this.id;
     }
     public void setLines(List<Line> lines){
@@ -33,7 +33,7 @@ public class Scene {
     public String getName(){
         return this.name;
     }
-    public void setOrdinal(int ordinal){
+    public void setOrdinal(long ordinal){
         this.ordinal = ordinal;
     }
     public long getOrdinal(){
